requirements.txt
torch
torchvision
Flask
flask-cors
Pillow